Restructure codebase
