--- crates/squareknot_core/src/lib.rs
@@ -1,11 +1,0 @@
-pub mod fast;
-pub mod graph;
-pub mod storage;
-pub mod undirected;
-
-pub use fast::*;
-pub use graph::*;
-pub use storage::*;
-pub use undirected::*;
-
-pub type SimpleGraph = UnGraph<AdjacencyList>;
