--- crates/squareknot_embedding/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "squareknot_embedding"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
