+++ crates/squareknot_graph/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "squareknot_core"
+name = "squareknot_graph"
 version = "0.1.0"
 edition = "2021"
