+++ crates/squareknot_graph/src/fast.rs
