+++ crates/squareknot_graph/src/graph.rs
