+++ crates/squareknot_graph/src/lib.rs
@@ -1,0 +1,15 @@
+//! This crate contains the core graph and storage types of `squareknot`.
+//! Additional graph and storage types may be defined outside of this crate.
+
+pub mod fast;
+pub mod graph;
+pub mod storage;
+pub mod undirected;
+
+pub use fast::*;
+pub use graph::*;
+pub use storage::*;
+pub use undirected::*;
+
+/// An alias for an undirected graph with an adjacency list.
+pub type SimpleGraph = UnGraph<AdjacencyList>;
