+++ crates/squareknot_graph/src/storage.rs
