+++ crates/squareknot_graph/src/storage/adjacency_list.rs
