+++ crates/squareknot_graph/src/undirected.rs
