+++ crates/squareknot_io/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2021"
 
 [dependencies]
 itertools = "0.14.0"
-squareknot_core = { path = "../squareknot_core" }
+squareknot_graph = { path = "../squareknot_graph" }
