+++ crates/squareknot_pathing/Cargo.toml
@@ -4,5 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-squareknot_core = { path = "../squareknot_core" }
+squareknot_graph = { path = "../squareknot_graph" }
 squareknot_traversal = { path = "../squareknot_traversal" }
