+++ crates/squareknot_planarity/src/lib.rs
@@ -1,0 +1,1 @@
+//! This crate contains functionality related to planarity testing.
