+++ crates/squareknot_traversal/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-squareknot_core = { path = "../squareknot_core" }
+squareknot_graph = { path = "../squareknot_graph" }
