+++ src/lib.rs
@@ -1,1 +1,3 @@
+//! Squareknot is a modular graph library, focused on both abstraction and performance.
+
 pub use squareknot_internal::*;
